# govconnect-dashboard/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GovConnect API
  description: |
    Complete API documentation for GovConnect microservices platform.
    
    ## Services
    - **Case Service** - Complaints & Service Requests management
    - **Channel Service** - WhatsApp Channel & Live Chat
    - **AI Service** - AI Orchestrator & Knowledge Base
    - **Notification Service** - Event-driven notifications
    
    ## Authentication
    - **Internal API Key**: `X-Internal-API-Key` header for service-to-service communication
    - **JWT Token**: `Authorization: Bearer <token>` for dashboard authentication
    - **Public APIs**: Endpoint publik tersedia untuk form publik (tanpa auth)
    
    ## Base URLs
    Layanan diakses langsung per service (tanpa gateway agregasi).
  version: 1.0.0
  contact:
    name: GovConnect Team
    url: https://govconnect.my.id
  license:
    name: MIT
servers:
  - url: https://case.govconnect.my.id
    description: Case Service (Production)
  - url: https://channel.govconnect.my.id
    description: Channel Service (Production)
  - url: https://ai.govconnect.my.id
    description: AI Service (Production)
...
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Internal API key for service-to-service communication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for dashboard authentication
  schemas:
...
    WebChatRequest:
      type: object
      required:
        - session_id
        - message
      properties:
        session_id:
          type: string
          example: "web_form_1234567890"
        message:
          type: string
        channel:
          type: string
          default: webchat
    WebChatResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
        intent:
          type: string
        processing_time_ms:
          type: integer
...
paths:
...
  /api/webchat:
    post:
      tags:
        - WebChat
      summary: Process webchat message
      description: Process a message from web chat interface
      operationId: processWebchatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebChatRequest'
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebChatResponse'
  /api/webchat/{session_id}:
    get:
      tags:
        - WebChat
      summary: Get session history
      operationId: getWebchatSessionHistory
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session history
    delete:
      tags:
        - WebChat
      summary: Clear session
      operationId: clearWebchatSession
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session cleared
...
  /api/webchat/{session_id}/poll:
    get:
      tags:
        - WebChat
      summary: Poll for messages
      description: Long polling for admin messages
      operationId: pollWebchatMessages
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: New messages if any

  # ============================================
  # AI SERVICE - Knowledge Vector
  # ============================================
...